Only filter by date if user selects one: #114
Set start/end date to the empty string and do not use them in the query
if the user did not choose a date.  Previously, if the user did not
filter requests by date we set the start and end dates to values that
were likely outside the request time period but had no relationship to
anything.  Now, we only filter requests by date if the user sets a start
and/or end date.

Also, fix #137.  It turned out that the default number of requests per
page was set to 500.  I changed it to 20, since I think that's more
likely to be what users actually want.  I also changed the default
sorting to be most to least recent creation date.

Note that the counting bug fixed in commit 6860441 is *not* fixed on
this branch, so even though the date filters work correctly, they have
no effect on the number of requests and pages shown at the bottom of the
admin page.