UI improvement for region checkboxes: #127
Move "No Region Found" out of general list of regions.  I might not have
done this in the most efficient way -- code review welcome!  Move
disabled regions to a separate column.

Add select/deselect all checkbox.  The behavior of this checkbox is not
correct yet.  Currently it is unchecked on click of "apply filters,"
even if all the enabled regions are (and remain) checked.  The next
commit will address this and other odd behaviors.

Add tests for the select/deselect all checkbox.