Preliminary #128 changes: save status change
Add a column for request status.  Save the status on change.  Currently
the save is being blocked by a permissions issue (that is, this will not
work as of this commit).  The save works if I disable all
permission-checking done with passport.  The next step is to resolve
that permissions problem.

Then, I'll look into displaying the status correctly.  UI might change
to a checkbox from the current select dropdown (since we only have two
statuses now).

Next: set all requests to have a status of 'new' when they are first
saved.