Pass CSRF token correctly for #128
Pass the CSRF token with the PUT request so that the new status is saved
without generating a 500 error.  Changing the select dropdown now saves
the request status to the db.

This requires exposing the CSRF token to the client in a hidden field.
I saw a lot of advice suggesting this method on the web, so I assume
it's okay from a security perspective (and the client already has access
to this token in a cookie anyway).  Review welcome, as ever.