Add ability to filter on status: #128
We didn't explicitly mention this in the issue, but I thought it would
be nice to have.  Users still can't sort on status, but they filter to
display only the new requests (for example).  In future, we might want
to turn this filter into a set of checkboxes, but for now we'll just
offer one value at a time.