Save requests incoming via SMS
Requests coming in via SMS are now saved.  As mentioned in the log
message for f3cdabb, requests coming in from SMS were for a time not
being saved.  I created a new "storeValues" function which I should be
able to use to fix the ongoing off-by-one error (the first text,
"welcome" and "what is your name" is currently being sent twice in a
row, for explicable reasons).

Next is still to fix that off-by-one error and do general code cleanup.
There are a lot of debugging calls in this commit -- I wanted to store
this state, where requests are saving.