Save raw SMS inputs: #121
Add new columns for raw address, zip, and phone inputs from SMS
requesters.  The name and email are already stored exactly as they
arrive, so we don't need to add specific columns for their raw forms.

Note that this change means the "request_object" cookie needs to be
created earlier, so that the raw information can be stored in it.

After this change, existing databases will need to run the migration
2016-add-raw-cols.sql.