Send help/info SMS response: #121
Respond to help/info texts appropriately.  Twilio told us that having a
generic response to those requests is a best practice (and in fact a
requirement for short code use).  Add a function that checks whether
we've received a request for info, and send the response accordingly.
Add tests for this case.  Send this response whether or not we've
completed the request, and make sure that we can start a new request
afterward.