Alert messages show up above the form #24
Added a jquery 'preventDefault' to prevent the submit button from
loading on to the next page so that it can let the user know if the
state, city, and zip code dont match. Appended an alert message that
shows the form has been completed. Appended the values of state, city,
and zip code below the alert message to use for on going tests.