Create new function and add to function saveRequest
Created checkZipMatch function that asks the database to run a query. The function returns the number of rows that matches the user input and  this will help check if the user's input exist in the database. Called the function in saveRequest to and added conditionals to bein testing each case.