Added ajax request (success)
Added an ajax request from the client to the server side and was
successfuly able to send back the response. The ajax request was made
to check to see if the city, state, and zip code matches. Many lines
of code are commented so that we can test the ajax call.