All required fields must be filled before submitting form #145
Tooltips will now work and the form will only succeed if all the
required fields are filled without having to change the language of
the page. The issue was that when the homepage loads, locale[1] did
not exist and it caused everything after it to break.I fixed this
issue by editing the conditional.