Add different error messages for each case #24
Created 3 more functions that checks if the number of rows matches the
different combinations in the database. Added conditionals to
exports.saveRequest to see which error message will appear when the
user enters in the incorrect combination of city, state, and zip code.
The error messages is not complete, need more tests.