Fixed case-insensitive issue in db-call #24
Created a function that capitalizes every word in the string and
applied it to the city parameter on every city-state-zip functions
that makes a db call. Prevoiusly, we would get an error if the user
filled in city field as "chicago" instead of "Chicago" (capitalized)
because the lower cased "chicago" does not exist in the database.