Got duplicate email check to work #149
I was able to get the first part of updating email and username to
work. I added "where" to the db call and removed
"_id: $ne:req.user.id" because column User._id does not exist. I also
added many console.logs to check which db calls are being made. The
issue is not fixed and is still in testing.