Exception Error in PatchUser #149
Workflow duplicate email and username is working without any issues
and updating the user and email is successful, however an exception
error appears on top of the form field/terminal. Console-logging the
'err' shows "[1]". I also console logged the user which appears that
user is undefined - that could be why the it raised an exception error
(not sure). There is another error I encountered in workflow-PatchAdmin,
"Cannot read property 'roles' which means that user is undefine. Will
look into this next time.