Duplicate email & user error message appears #149
An error message will now appear below the field if a duplicate email
or username already exists in the database. An issue that is still
ocurring is that the user must update both the username and email in
order for the update to work. Removed extra code from drywall that is
not needed and some console logs from previous tests. Edited the
functions in duplicateEmailCheck and duplicateUserCheck because there
should only be one argument instead of 2 (err, user).