Add delivery receipt callback to SMS: #121
As suggested by @gregbaugues, add a delivery receipt check to our
SMS-sending code.  At the moment, it doesn't look like I've set it up
correctly, because nothing is being POSTed to the statusCallback
endpoint.  So, this commit is purely for reference from the discussion
in the issue.

Once we get the status in this endpoint, I'll resend the text when we
see that it is undelivered.  That will (with luck) ameliorate the
problem where some people never receive a certain text, despite our logs
showing that it was sent.