Remove patchAdmin call from patchUser #149
Removed patchAdmin call because patchAdmin, patchAccount and
populateRoles have an attribute that does not apply to user and account.
For example, we get an error for `user.roles.admin` if patchAdmin is
called because user has no 'role' attribute. This part of the code is
probably left over from Drywall.