Briefly delay status check: #121
Check the status of the most recent text message a few seconds after
sending that text message, so that we're checking the status of the text
we just sent, instead of the immediately previous text.  Add logging so
that we know when texts were sent correctly.

Next is to handle the case where the text isn't delivered.  I could just
resend it, but I worry about causing an infinite loop.