Fix CSV exports when no region selected #160
Previously when no region was selected, the exported CSV file
contained all allowed data. Remove the codeblock that caused this, in
order to keep displayed exported data congruent with displayed data,
which is blank in this case. I don't understand exactly why this was
there originally, but @murraymel and I discussed it with @cecilia-
donnelly and @kfogel we agreed that a blank CSV is expected.