Start handling failed SMS: #121
Test the status of the a sent SMS and resend it if it has failed.  Note
that we're calling the Twilio API to get the information about the sent
text, which may not be the most efficient way to handle this. In
forthcoming commits, I'll handle the case where a text fails multiple
times.  For now I've included a max number of resends, but we need to
actually track the failures across attempts.