Correctly resend texts: #121
Use Twilio API correctly to resend a text.  Since we're not technically
responding to an incoming text when we're resending a failed SMS, we
need to send it as an outgoing message instead of as a response.

Note that the statusCallback here needs to be a full path, not a
relative path within the application.