Complete fix to #160
Overwrite serialize function in all queries from the client, including
CSV.  Reinstate default case on first page load (removed in 83d4d5e).
Clarify some server-side code.

For more detail:

The change in 83d4d5e fixed #160, but also had the following unintended
consequence: after it, when an admin user logged into the request list,
s/he would see no results on the front page until s/he applied a filter
or clicked "clear filters."

Think of it this way: when a request for this page hits the server, it
either has or does not have a query.  The only time it doesn't have a
query is on first page load, when we expect the default results to show
up (all results in regions that the user has access to).  In the case
where the request has a query, it may or may not include any regions.
If it does not include any regions, then #160 shows up for requests that
also asked for a response in CSV format.  Here are the cases, restated:

1. No query (usually first page load)
2. Query (either CSV or web filtering)
    2a. Query with one or more regions
    2b. Query with no regions

Commit 83d4d5e removed a block of code and altered the no-query case: it
changed the default response from "all results in all the regions to
which the user has access" to "no results in any region."  After
83d4d5e, everything worked in cases 2a and 2b but case 1 was broken.

Of course, restoring the codeblock broke case 2b for CSV exports again,
as described in #160.  So, I also applied a new patch: in the
client-side javascript, construct the CSV query consistently with the
way the filtering queries are constructed.  For both kinds of queries
(CSV and web filtering), serializing the form with no regions checked
should pass an empty region array.  The generic "serialize" function
does *not* have this behavior: instead, if no regions are checked then
the query simply doesn't include a region element.  Unfortunately, with
no region element passed the server thinks we're in case 1, and by
default displays the results in all regions that the logged-in user has
access to, just like on first page load.

To work around this, I had added an empty region array for case 2b in
commit 44f3219.  I only did this for web filtering queries, though, and
neglected to do the same for CSV queries.  *That* oversight yielded the
bug described in #160.  To fix that bug, in this commit I overwrote
"serialize" to include adding the empty region array in the case where
the user has unchecked all the regions, and used the new "serializeForm"
function when constructing queries for both CSV and non-CSV requests to
the server.