Offer alternative kiosk route: #202
Create a "kiosk" route.  Conditionally enable the kiosk's refresh
function based on whether the request came from the kiosk route or the
regular homepage.  Pass the origin of the request to both thank you and
sorry pages, to make this possible.

This means that we can merge this into master, not alter the workings of
the homepage, and have a functional kiosk all on the same instance.