Send new request to external endpoint: #196
As requested by @MisterJames, we can now send a new request to some
external endpoint (which will be determined in the config file).  I need
to check whether the request is in an active region first and *then*
send it.  I should also not try to do this if no external endpoint has
been chosen.  This is an intermediate step, in other words.  Once I
clean up those loose ends I still need to add proper auth.