Start #205 to point @adamtegen to the right spot
Add function to handle saving partial smoke alarm requests.  This is
completely untested (it's essentially half of a commit), but should help
@adamtegen find the right area of code to work in for #205.

General notes: currently we're saving the data from an SMS conversation
into the `priorities` cookie, but we want to store it in the database
instead, so that it isn't lost if the user stops responding to texts.  I
think the `req.cookies.priorities` cookie can be removed entirely as
part of this process.

If you need help setting up a test Twilio account, you might be able to
ask @kwhinnery.