Alter availability language for newly active KSNE
Since the KSNE region is active now, reflect that in the front-page
language about who can use this app.  Explicitly change the text in both
our English and Spanish files, and point to that text with a more
meaningful key in "index.jade," rather than the English version of the
text.  If we move to using keys in the template files instead of text
then it'll be easier to remember to change the text in our locale files.