Restricting user management to superusers: #138
Allow superusers to designate others as "admins"/"superusers."

- Link users to admin groups directly (via a migration)
- Add new column to User table, to link to the adminGroup table.
- Add new "role" of superuser
- Limit user* routes to users that have admin privileges
- Replace use of "isMemberOf" with "canPlayRoleOf" (I didn't know how to
  set a user as a member of "root")
- Multiple users can be linked to the same adminGroup, so remove
  duplicateLinkCheck.  Similarly, don't 'patchAdmin', since we're
  bypassing the 'Admin' db table entirely
- Fix db connections and use then/catch properly

Some more context: Previously, "Admins" were members of "AdminGroups".
The User:Admin relationship was 1:1.  So, a User object would be linked
to an Admin object, which would be a member of an AdminGroup.
Privileges are determined by the AdminGroup that a User was linked to
via the Admin.  I couldn't see why we needed the "Admin" layer at all.
I removed it (from this part of the app, not the entire thing) and have
now, with migration `20170424-admingroup.sql` linked Users directly to
AdminGroups.

Front end:
- Fix the redirection URLs that are used after each change is made to
  the user.
- Remove old Admin/Account inputs from the page
- Create a new checkbox, for turning on admin privileges for a user
  (this is a WIP -- currently it only turns *on*, not *off*)

Next steps:

- Manage the magic numbers that I've put in a couple places (grep for
  "TODO" or even "magic" to find them)
- Turn *off* admin privileges for a user, from the user details page
- Show whether a user is an admin or not, using the checkbox (currently
  it doesn't stay checked after a user is set as an admin)