Add model and handler for duplicates
This adds a RequestDuplicate model with requestId as its only field
other than the defaults, as well as several methods combined in
saveOrDuplicateRequest that check if a request exists, update the
updateAt column and create a RequestDuplicate if it does, and
otherwise creates a new record as normal. Querying is based on mostly
user input-based fields with a case-insensitive lookup that is flexible
for phone number formatting. Also updates Request to use more standard
Sequelize syntax.