Fix basic save and remove roles from user details.
Updating users wasn't working due to (potentially) old code from
sequelize.  As an example, the findByIdAndUpdate, upon a quick internet
search, function seems to be used by mongoose, and was not working with
sequelize.

While in there, removed the roles part of the User interface because
we aren't using that, and it was getting in the way of testing.

Issue #263: Admin users and User Administration screen