Add SiteAdmin to edit and list interfaces.
Replaces the unused isActive flag and replaces with the siteAdmin flag,
both in the edit user screen, and in the user list screen.  Of note is
that the isActive may be reenabled in a later commit, but will have to
be reworked to use a checkbox like the siteAdmin flag is added here.

Issue #263: Admin users and User Administration screen