Fix searching users by username.
The code underlying wasn't working with sequelize, and not fully fleshed
out.  At the same time, removed the other search options, because the
likelihood of someone wanting to search for classes of users is very low
given the expected user count.  Optimized for the common, and most
likely unique, case of looking for a specific user by their username.

Issue #263: Admin users and User Administration screen