Remove unnecessary contact parameter
constructFinalText took a contact parameter that was unused.  Because
we're updating the region contacts to be a list rather than a single
person, constructFinalText would have needed to be updated to represent
that change in some way.  However, because the parameter wasn't used,
we sidestep that design issue and just remove it.

Related to issue #264: one contact per region