Make user list active/admin columns checkboxes.
This required not only changing the checkboxes, but also updating the id
on the model so that Backbone would know how to update a specific user.
The decision was made to reuse the current edit user path, with a
special identifier let the server know that it's only partial
information coming in, so as to not clutter up the endpoint namespace.

Issue #263: Admin users and User Administration screen