Change user edit to support user add.
Instead of the old way of adding a user, where you set a username, then
edit that user, present the user details screen with empty fields.  Then
when adding the user, do a verification of all the fields so that the
user can be added in one step.

Issue #263: Admin users and User Administration screen