Change SelectedCountes to be related to chapter
Counties no longer related directly to region to allow more fine grained
control of which section of redcross they relate to.  They now link to
chapters, which should link to regions.  This commit is only for the
migration, with further commits related to the code to follow.

Issue #277: Add County, Chapter Code and Chapter Name